feat: add mock translation vertical slice

+++ crates/lexift-core/src/domain/geometry.rs
@@ -1,5 +1,5 @@
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
+    pub x: i32,
+    pub y: i32,
 }
